Move filtering out of list component
